# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogEntryResponseDto": {
        "properties": {
          "catalog": {
            "enum": [
              "division",
              "gender_format",
              "position"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "referenceCount": {
            "type": "number"
          },
          "sortOrder": {
            "type": "number"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "type": "string"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "nullable": true,
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "teamId",
          "catalog",
          "key",
          "label",
          "sortOrder",
          "metadata",
          "referenceCount",
          "status",
          "createdBy",
          "updatedBy",
          "createdAt",
          "updatedAt",
          "version"
        ],
        "type": "object"
      },
...
      "MembershipResponseDto": {
        "properties": {
          "anonymizedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "nullable": true,
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "leftAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "seasonId": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "invited",
              "active",
              "inactive",
              "suspended",
              "left",
              "archived",
              "anonymized"
            ],
            "type": "string"
          },
          "statusEffectiveAt": {
            "format": "date-time",
            "type": "string"
          },
          "statusReason": {
            "nullable": true,
            "type": "string"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "nullable": true,
            "type": "string"
          },
          "userId": {
            "nullable": true,
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "teamId",
          "seasonId",
          "userId",
          "status",
          "statusReason",
          "statusEffectiveAt",
          "joinedAt",
          "leftAt",
          "anonymizedAt",
          "createdBy",
          "updatedBy",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "version"
        ],
        "type": "object"
      },
...
      "ScheduleResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "nullable": true,
            "type": "string"
          },
          "defaultCapacity": {
            "nullable": true,
            "type": "number"
          },
          "defaultField": {
            "nullable": true,
            "type": "string"
          },
          "defaultVenueId": {
            "nullable": true,
            "type": "string"
          },
          "durationMinutes": {
            "type": "number"
          },
          "exceptions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "frequency": {
            "enum": [
              "weekly",
              "one_off"
            ],
            "type": "string"
          },
          "generationStart": {
            "type": "string"
          },
          "generationUntil": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "intervalWeeks": {
            "type": "number"
          },
          "meetOffsetMinutes": {
            "nullable": true,
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "organizerUserId": {
            "nullable": true,
            "type": "string"
          },
          "rsvpCutoffMinutes": {
            "nullable": true,
            "type": "number"
          },
          "seasonId": {
            "nullable": true,
            "type": "string"
          },
          "sessionType": {
            "type": "string"
          },
          "startTimeLocal": {
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "type": "string"
          },
          "teamId": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "nullable": true,
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "visibility": {
            "enum": [
              "team",
              "coaches",
              "public"
            ],
            "type": "string"
          },
          "weekdays": {
            "items": {
              "type": "number"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "teamId",
          "seasonId",
          "name",
          "sessionType",
          "timezone",
          "frequency",
          "intervalWeeks",
          "weekdays",
          "startTimeLocal",
          "durationMinutes",
          "meetOffsetMinutes",
          "rsvpCutoffMinutes",
          "defaultVenueId",
          "defaultField",
          "defaultCapacity",
          "visibility",
          "organizerUserId",
          "notes",
          "generationStart",
          "generationUntil",
          "exceptions",
          "status",
          "createdBy",
          "updatedBy",
          "createdAt",
          "updatedAt",
          "version"
        ],
        "type": "object"
      },
...
      "TeamResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "logoMediaKey": {
            "nullable": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "primaryColor": {
            "nullable": true,
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "nullable": true,
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "locale",
          "timezone",
          "primaryColor",
          "logoMediaKey",
          "status",
          "createdBy",
          "updatedBy",
          "createdAt",
          "updatedAt",
          "version"
        ],
        "type": "object"
      },
      "TransitionDto": {
        "properties": {
          "effectiveAt": {
            "format": "date-time",
            "type": "string"
          },
          "reason": {
            "maxLength": 512,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateScheduleDto": {
        "properties": {
          "defaultCapacity": {
            "maximum": 10000,
            "minimum": 0,
            "type": "number"
          },
          "defaultField": {
            "maxLength": 120,
            "type": "string"
          },
          "defaultVenueId": {
            "format": "uuid",
            "type": "string"
          },
          "durationMinutes": {
            "maximum": 1440,
            "minimum": 1,
            "type": "number"
          },
          "exceptions": {
            "items": {
              "type": "string"
            },
            "maxItems": 366,
            "type": "array"
          },
          "expectedVersion": {
            "minimum": 1,
            "type": "number"
          },
          "frequency": {
            "enum": [
              "weekly",
              "one_off"
            ],
            "type": "string"
          },
          "generationStart": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "type": "string"
          },
          "generationUntil": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "type": "string"
          },
          "intervalWeeks": {
            "maximum": 8,
            "minimum": 1,
            "type": "number"
          },
          "meetOffsetMinutes": {
            "maximum": 1440,
            "minimum": 0,
            "type": "number"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "notes": {
            "maxLength": 2000,
            "type": "string"
          },
          "organizerUserId": {
            "format": "uuid",
            "type": "string"
          },
          "rsvpCutoffMinutes": {
            "maximum": 1440,
            "minimum": 0,
            "type": "number"
          },
          "seasonId": {
            "format": "uuid",
            "type": "string"
          },
          "sessionType": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "startTimeLocal": {
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "type": "string"
          },
          "timezone": {
            "maxLength": 64,
            "type": "string"
          },
          "visibility": {
            "enum": [
              "team",
              "coaches",
              "public"
            ],
            "type": "string"
          },
          "weekdays": {
            "items": {
              "type": "number"
            },
            "maxItems": 7,
            "type": "array"
          }
        },
        "required": [
          "name",
          "sessionType",
          "frequency",
          "startTimeLocal",
          "durationMinutes",
          "generationStart",
          "generationUntil",
          "status",
          "expectedVersion"
        ],
        "type": "object"
      },
...
      "UpdateTeamDto": {
        "properties": {
          "expectedVersion": {
            "minimum": 1,
            "type": "number"
          },
          "locale": {
            "maxLength": 16,
            "type": "string"
          },
          "logoMediaKey": {
            "maxLength": 256,
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "primaryColor": {
            "maxLength": 32,
            "type": "string"
          },
          "timezone": {
            "maxLength": 64,
            "type": "string"
          }
        },
        "required": [
          "name",
          "expectedVersion"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "HTTP API for this NestJS service",
    "title": "Service API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/teams/{teamId}": {
      "delete": {
        "operationId": "Teams.archive",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            },
            "description": "Team archived"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Archive a team",
        "tags": [
          "teams"
        ]
      },
      "get": {
        "operationId": "Teams.get",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            },
            "description": "Team"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get a team by id",
        "tags": [
          "teams"
        ]
      },
      "patch": {
        "operationId": "Teams.update",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            },
            "description": "Team updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Update a team",
        "tags": [
          "teams"
        ]
      }
    },
...
    "/teams/{teamId}/catalog-entries/{entryId}": {
      "delete": {
        "operationId": "Catalogs.archive",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntryResponseDto"
                }
              }
            },
            "description": "Catalog entry archived"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Archive a reference catalog entry",
        "tags": [
          "teams"
        ]
      }
    },
...
    "/teams/{teamId}/members/{membershipId}/archive": {
      "post": {
        "operationId": "Members.archive",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponseDto"
                }
              }
            },
            "description": "Member archived"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Archive a member",
        "tags": [
          "members"
        ]
      }
    },
...
    "/teams/{teamId}/practice-schedules/{scheduleId}": {
      "delete": {
        "operationId": "PracticeSchedules.archive",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scheduleId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            },
            "description": "Schedule archived"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Archive a practice schedule",
        "tags": [
          "practices"
        ]
      },
      "get": {
        "operationId": "PracticeSchedules.get",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scheduleId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            },
            "description": "Schedule"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get a practice schedule",
        "tags": [
          "practices"
        ]
      },
      "patch": {
        "operationId": "PracticeSchedules.update",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scheduleId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            },
            "description": "Schedule updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Update a practice schedule template",
        "tags": [
          "practices"
        ]
      }
    },
...
    "/teams/{teamId}/seasons/{seasonId}": {
      "delete": {
        "operationId": "Seasons.archive",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "seasonId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponseDto"
                }
              }
            },
            "description": "Season archived"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Archive a season",
        "tags": [
          "teams"
        ]
      },
      "patch": {
        "operationId": "Seasons.update",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "seasonId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSeasonDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponseDto"
                }
              }
            },
            "description": "Season updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Update a season",
        "tags": [
          "teams"
        ]
      }
    },
...
    "/teams/{teamId}/venues/{venueId}": {
      "delete": {
        "operationId": "Venues.archive",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "venueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueResponseDto"
                }
              }
            },
            "description": "Venue archived"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Archive a venue",
        "tags": [
          "teams"
        ]
      },
      "patch": {
        "operationId": "Venues.update",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "venueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueResponseDto"
                }
              }
            },
            "description": "Venue updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Update a venue",
        "tags": [
          "teams"
        ]
      }
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
  "servers": [],
...
}